x12Batch object

x12Batch object of four AirPassengers series with paramters and output objects

data(AirPassengersX12Batch)

Examples

data(AirPassengersX12Batch)
summary(AirPassengersX12Batch)
#> -----------------------------------------------------------------------------------
#> --------------------------   Series_1   ------------------------------------
#> -----------------------------------------------------------------------------------
#> 
#> 	Time Series
#> 
#> Frequency: 12 
#> Span: 1st month,1949 to 12th month,1960 
#> 
#> 	Model Definition
#> 
#> ARIMA Model: (1,1,0)(0,1,1) 
#> Model Span: 1st month,1949 to 12th month,1960 
#> Transformation: Automatic selection : Log(y) 
#> Regression Model: none 
#> 
#> 	Outlier Detection
#> 
#> Outlier Span: 1st month,1949 to 12th month,1960 
#> Critical |t| for outliers:	
#> aocrit1 aocrit2 lscrit1 lscrit2 tccrit1 tccrit2 
#>  "3.89"     "*"  "3.89"     "*"  "3.89"     "*" 
#> Total Number of Outliers: 0 
#> Automatically Identified Outliers: 0 
#> 
#> 	Seasonal Adjustment
#> 
#> Identifiable Seasonality: yes 
#> Seasonal Peaks: rsd 
#> Trading Day Peaks: sa irr 
#> Overall Index of Quality of SA
#> (Acceptance Region from 0 to 1)
#> Q: 0.26 
#> Number of M statistics outside the limits: 0 
#> 
#> SA decomposition: multiplicative 
#> Seasonal moving average used for the final iteration: 
#> 3x3 (Based on the size of the global moving seasonality ratio (msr))
#> Moving average used to estimate the final trend-cycle: 9-term Henderson filter
#> -----------------------------------------------------------------------------------
#> --------------------------   Series_2   ------------------------------------
#> -----------------------------------------------------------------------------------
#> 
#> 	Time Series
#> 
#> Frequency: 12 
#> Span: 1st month,1949 to 12th month,1960 
#> 
#> 	Model Definition
#> 
#> ARIMA Model: (0 1 1)(0 1 1) (Automatic Model Choice)
#> Model Span: 1st month,1949 to 12th month,1960 
#> Transformation: Automatic selection : Log(y) 
#> Regression Model: Automatically Identified Outliers 
#> 
#> 	Outlier Detection
#> 
#> Outlier Span: 1st month,1949 to 12th month,1960 
#> Critical |t| for outliers:	
#> aocrit1 aocrit2  lscrit  tccrit 
#>  "3.89"     "*"   "3.5"   "2.5" 
#> Total Number of Outliers: 6 
#> Automatically Identified Outliers: 6 
#> 
#> 	Regression Model
#>                 variable   coef stderr   tval
#> 1 autooutlier_TC1951.May  0.078  0.023  3.321
#> 2 autooutlier_TC1951.Jun -0.099  0.024 -4.174
#> 3 autooutlier_TC1952.Mar -0.083  0.023 -3.610
#> 4 autooutlier_LS1953.Jun -0.090  0.023 -3.851
#> 5 autooutlier_TC1954.Feb -0.075  0.023 -3.243
#> 6 autooutlier_AO1960.Mar -0.104  0.024 -4.270
#> 
#> 	Seasonal Adjustment
#> 
#> Identifiable Seasonality: yes 
#> Seasonal Peaks: rsd 
#> Trading Day Peaks: rsd 
#> Overall Index of Quality of SA
#> (Acceptance Region from 0 to 1)
#> Q: 0.2 
#> Number of M statistics outside the limits: 0 
#> 
#> SA decomposition: multiplicative 
#> Seasonal moving average used for the final iteration: 
#> 3x3 (Based on the size of the global moving seasonality ratio (msr))
#> Moving average used to estimate the final trend-cycle: 9-term Henderson filter
#> -----------------------------------------------------------------------------------
#> --------------------------   Series_3   ------------------------------------
#> -----------------------------------------------------------------------------------
#> 
#> 	Time Series
#> 
#> Frequency: 12 
#> Span: 1st month,1949 to 12th month,1960 
#> 
#> 	Model Definition
#> 
#> ARIMA Model: (0 1 1)(0 1 1) (Automatic Model Choice)
#> Model Span: 1st month,1949 to 12th month,1960 
#> Transformation: Automatic selection : Log(y) 
#> Regression Model: none 
#> 
#> 	Outlier Detection
#> 
#> Outlier Span: 1st month,1949 to 12th month,1960 
#> Critical |t| for outliers:	
#> aocrit1 aocrit2 lscrit1 lscrit2 tccrit1 tccrit2 
#>  "3.89"     "*"  "3.89"     "*"  "3.89"     "*" 
#> Total Number of Outliers: 0 
#> Automatically Identified Outliers: 0 
#> 
#> 	Seasonal Adjustment
#> 
#> Identifiable Seasonality: yes 
#> Seasonal Peaks: rsd 
#> Trading Day Peaks: sa irr 
#> Overall Index of Quality of SA
#> (Acceptance Region from 0 to 1)
#> Q: 0.26 
#> Number of M statistics outside the limits: 0 
#> 
#> SA decomposition: multiplicative 
#> Seasonal moving average used for the final iteration: 
#> 3x3 (Based on the size of the global moving seasonality ratio (msr))
#> Moving average used to estimate the final trend-cycle: 9-term Henderson filter
#> -----------------------------------------------------------------------------------
#> --------------------------   Series_4   ------------------------------------
#> -----------------------------------------------------------------------------------
#> 
#> 	Time Series
#> 
#> Frequency: 12 
#> Span: 1st month,1949 to 12th month,1960 
#> 
#> 	Model Definition
#> 
#> ARIMA Model: (0 1 1)(0 1 1) (Automatic Model Choice)
#> Model Span: 1st month,1949 to 12th month,1960 
#> Transformation: Automatic selection : Log(y) 
#> Regression Model: none 
#> 
#> 	Outlier Detection
#> 
#> Outlier Span: 1st month,1949 to 12th month,1960 
#> Critical |t| for outliers:	
#> aocrit lscrit tccrit 
#>    3.5    3.5    3.5 
#> Total Number of Outliers: 0 
#> Automatically Identified Outliers: 0 
#> 
#> 	Seasonal Adjustment
#> 
#> Identifiable Seasonality: yes 
#> Seasonal Peaks: rsd 
#> Trading Day Peaks: sa irr 
#> Overall Index of Quality of SA
#> (Acceptance Region from 0 to 1)
#> Q: 0.26 
#> Number of M statistics outside the limits: 0 
#> 
#> SA decomposition: multiplicative 
#> Seasonal moving average used for the final iteration: 
#> 3x3 (Based on the size of the global moving seasonality ratio (msr))
#> Moving average used to estimate the final trend-cycle: 9-term Henderson filter