~~ Methods for Function plotRsdAcf in Package x12 ~~

Plot of the (partial) autocorrelations of the (squared) residuals with function plotRsdAcf in package x12.

# S4 method for class 'x12Output'
plotRsdAcf(x, which="acf",
        xlab="Lag", ylab="ACF",
        main="default", col_acf="darkgrey", lwd_acf=4,
        col_ci="blue", lt_ci=2, ylim="default", ...)
# S4 method for class 'x12Single'
plotRsdAcf(x, which="acf",
        xlab="Lag", ylab="ACF",
        main="default", col_acf="darkgrey", lwd_acf=4,
        col_ci="blue", lt_ci=2, ylim="default", ...)

Methods

signature(x = "x12Output")

signature(x = "x12Single")

Arguments

x

object of class x12Output-class or x12Single-class.

which

character specifying the type of autocorrelation of the residuals that should be plotted, i.e. the autocorrelations or partial autocorrelations of the residuals or the autocorrelations of the squared residuals ("acf", "pacf", "acf2").

xlab

label of the x-axis.

ylab

label of the y-axis.

main

plot title.

col_acf

color of the autocorrelations.

lwd_acf

line width of the autocorrelations.

col_ci

color of the +- 2 standard error limits.

lt_ci

line type of the +- 2 standard error limits.

ylim

range of the y-axis.

...

ignored.

Author

Alexander Kowarik, Angelika Meraner

See also

Examples

if (FALSE) { # \dontrun{
s <- new("x12Single",ts=AirPassengers,tsName="air")
s <- setP(s,list(estimate=TRUE,regression.variables="AO1950.1",outlier.types="all",
  outlier.critical=list(LS=3.5,TC=2.5),backcast_years=1/2))
s <- x12(s)
#w/o outliers
plot(s@x12Output,sa=TRUE,trend=TRUE,original=FALSE)
plot(s)
#with (all) outliers
plot(s,showAllout=TRUE,sa=TRUE,trend=TRUE,log_transform=TRUE,lwd_out=1,pch_ao=4)
plot(s,showAllout=TRUE,sa=TRUE,trend=TRUE,original=FALSE,showAlloutLines=TRUE,
  col_tc="purple")#,log_transform=TRUE)#,lwd_out=3)
#with showOut
plot(s,showOut="AO1960.Jun",sa=FALSE,trend=FALSE,annComp=TRUE,log_transform=TRUE)
plot(s,showOut="AO1958.Mar",sa=TRUE,trend=TRUE,annComp=TRUE,annCompTrend=FALSE)
plot(s,showOut="AO1950.Jun",annComp=FALSE,cex_out=3,pch_ao=19,col_ao="orange")
plot(s,showOut="TC1954.Feb")
plot(s,showOut="TC1954.Feb",col_tc="green3")

#w/o legend
plot(s,showAllout=TRUE,plot_legend=FALSE)
plot(s,plot_legend=FALSE)
plot(s,showOut="AO1950.1",plot_legend=FALSE,lwd_out=2,col_ao="purple")
plot(s,showOut="TC1954.Feb",col_tc="orange",col_ao="magenta",plot_legend=FALSE)
plot(s,showOut="AO1950.1",col_tc="orange",col_ao="magenta",plot_legend=FALSE)

#Forecasts & Backcasts
plot(s,forecast=TRUE)
plot(s,backcast=TRUE,showLine=TRUE)
plot(s,backcast=TRUE,forecast=TRUE,showCI=FALSE)
plot(s,forecast=TRUE,points_fc=TRUE,col_fc="purple",lty_fc=2,lty_original=3,lwd_fc=0.9,
  lwd_ci=2)
plot(s,sa=TRUE,plot_legend=FALSE)

#Seasonal Factors and SI Ratios
plotSeasFac(s)
#Spectra
plotSpec(s)
plotSpec(s,highlight=FALSE)
#Autocorrelations of the Residuals
plotRsdAcf(s)
plotRsdAcf(s,col_acf="black",lwd_acf=1)
} # }