~~ Methods for Function plotSpec in Package x12 ~~

Spectral plots with function plotSpec in package x12.

Methods

signature(x = "x12Output",which="sa", xlab="Frequency",ylab="Decibels", main="Spectrum", col_bar="darkgrey",col_seasonal="red",col_td="blue", lwd_bar=4,lwd_seasonal=4,lwd_td=4,plot_legend=TRUE,...)

signature(x = "x12Single",which="sa", xlab="Frequency",ylab="Decibels", main="Spectrum", col_bar="darkgrey",col_seasonal="red",col_td="blue", lwd_bar=4,lwd_seasonal=4,lwd_td=4,plot_legend=TRUE,...)

signature(x = "spectrum",frequency, xlab="Frequency",ylab="Decibels", main="Spectrum", col_bar="darkgrey",col_seasonal="red",col_td="blue", lwd_bar=4,lwd_seasonal=4,lwd_td=4,plot_legend=TRUE,...)

Arguments

x

an object of class x12Output-class, x12Single-class or spectrum-class.

which

a string defining the executable of the editor to use ("sa" for the Spectrum of the Seasonally Adjusted Series, "original" for the Spectrum of the Original Series, "irregular" for the Spectrum of the Irregular Series and "residuals" for the Spectrum of the RegARIMA Residuals).

frequency

frequency of the time series (has to be specified for objects of class "spectrum" only).

xlab

label of the x-axis.

ylab

label of the y-axis.

main

plot title.

col_bar

color of bars.

col_seasonal

color of seasonal frequencies.

col_td

color of trading day frequencies.

lwd_bar

line width of bars.

lwd_seasonal

line width of seasonal frequencies.

lwd_td

line width of trading day frequencies.

plot_legend

logical specifying if a legend should be plotted.

Author

Alexander Kowarik, Angelika Meraner

See also

Examples

if (FALSE) {
s <- new("x12Single",ts=AirPassengers,tsName="air")
s <- setP(s,list(estimate=TRUE,regression.variables="AO1950.1",outlier.types="all",
  outlier.critical=list(LS=3.5,TC=2.5),backcast_years=1/2))
s <- x12(s)
#w/o outliers
plot(s@x12Output,sa=TRUE,trend=TRUE,original=FALSE)
plot(s)
#with (all) outliers
plot(s,showAllout=TRUE,sa=TRUE,trend=TRUE,log_transform=TRUE,lwd_out=1,pch_ao=4)
plot(s,showAllout=TRUE,sa=TRUE,trend=TRUE,original=FALSE,showAlloutLines=TRUE,
  col_tc="purple")#,log_transform=TRUE)#,lwd_out=3)
#with showOut
plot(s,showOut="AO1960.Jun",sa=FALSE,trend=FALSE,annComp=TRUE,log_transform=TRUE)
plot(s,showOut="AO1958.Mar",sa=TRUE,trend=TRUE,annComp=TRUE,annCompTrend=FALSE)
plot(s,showOut="AO1950.Jun",annComp=FALSE,cex_out=3,pch_ao=19,col_ao="orange")
plot(s,showOut="TC1954.Feb")
plot(s,showOut="TC1954.Feb",col_tc="green3")

#w/o legend
plot(s,showAllout=TRUE,plot_legend=FALSE)
plot(s,plot_legend=FALSE)
plot(s,showOut="AO1950.1",plot_legend=FALSE,lwd_out=2,col_ao="purple")
plot(s,showOut="TC1954.Feb",col_tc="orange",col_ao="magenta",plot_legend=FALSE)
plot(s,showOut="AO1950.1",col_tc="orange",col_ao="magenta",plot_legend=FALSE)

#Forecasts & Backcasts
plot(s,forecast=TRUE)
plot(s,backcast=TRUE,showLine=TRUE)
plot(s,backcast=TRUE,forecast=TRUE,showCI=FALSE)
plot(s,forecast=TRUE,points_fc=TRUE,col_fc="purple",lty_fc=2,lty_original=3,
  lwd_fc=0.9,lwd_ci=2)
plot(s,sa=TRUE,plot_legend=FALSE)

#Seasonal Factors and SI Ratios
plotSeasFac(s)
#Spectra
plotSpec(s)
plotSpec(s,highlight=FALSE)
#Autocorrelations of the Residuals
plotRsdAcf(s)
plotRsdAcf(s,col_acf="black",lwd_acf=1)
}