~~ Methods for Function prev and cleanArchive in Package x12 ~~

Function prev in package x12 reverts to previous parameter settings and output.
Function cleanHistory resets x12OldParameter and x12OldOutput.

# S4 method for x12Single
prev(object,n=NULL)
# S4 method for x12Batch
prev(object,index=NULL,n=NULL)
# S4 method for x12Single
cleanHistory(object)
# S4 method for x12Batch
cleanHistory(object,index=NULL)

Methods

signature(object = "x12Single")

signature(object = "x12Batch")

Arguments

object

object of class x12Single-class or x12Batch-class.

n

index corresponding to a previous run.

index

index corresponding to (an) object(s) of class "x12Single".

See also

Note

cleanHistory is deprecated and cleanArchive should be used instead.

Author

Alexander Kowarik

Examples

data(AirPassengersX12)
summary(AirPassengersX12)
#> --------------------------   AirPassengers   ------------------------------------
#> -----------------------------------------------------------------------------------
#> 
#> 	Time Series
#> 
#> Frequency: 12 
#> Span: 1st month,1949 to 12th month,1960 
#> 
#> 	Model Definition
#> 
#> ARIMA Model: (1,1,0)(0,1,1) 
#> Model Span: 1st month,1949 to 12th month,1960 
#> Transformation: Automatic selection : Log(y) 
#> Regression Model: none 
#> 
#> 	Outlier Detection
#> 
#> Outlier Span: 1st month,1949 to 12th month,1960 
#> Critical |t| for outliers:	
#> aocrit1 aocrit2 lscrit1 lscrit2 tccrit1 tccrit2 
#>  "3.89"     "*"  "3.89"     "*"  "3.89"     "*" 
#> Total Number of Outliers: 0 
#> Automatically Identified Outliers: 0 
#> 
#> 	Seasonal Adjustment
#> 
#> Identifiable Seasonality: yes 
#> Seasonal Peaks: rsd 
#> Trading Day Peaks: sa irr 
#> Overall Index of Quality of SA
#> (Acceptance Region from 0 to 1)
#> Q: 0.26 
#> Number of M statistics outside the limits: 0 
#> 
#> SA decomposition: multiplicative 
#> Seasonal moving average used for the final iteration: 
#> 3x3 (Based on the size of the global moving seasonality ratio (msr))
#> Moving average used to estimate the final trend-cycle: 9-term Henderson filter
# a maximum of 10 previous x12 runs are added to the summary
summary(AirPassengersX12,oldOutput=10)
#> --------------------------   AirPassengers   ------------------------------------
#> -----------------------------------------------------------------------------------
#> 
#> 	Time Series
#> 
#> Frequency: 12 
#> Span: 1st month,1949 to 12th month,1960 
#> 
#> 	Model Definition
#> 
#> ARIMA Model: (1,1,0)(0,1,1) 
#> Model Span: 1st month,1949 to 12th month,1960 
#> Transformation: Automatic selection : Log(y) 
#> Regression Model: none 
#> 
#> 	Outlier Detection
#> 
#> Outlier Span: 1st month,1949 to 12th month,1960 
#> Critical |t| for outliers:	
#> aocrit1 aocrit2 lscrit1 lscrit2 tccrit1 tccrit2 
#>  "3.89"     "*"  "3.89"     "*"  "3.89"     "*" 
#> Total Number of Outliers: 0 
#> Automatically Identified Outliers: 0 
#> 
#> 	Seasonal Adjustment
#> 
#> Identifiable Seasonality: yes 
#> Seasonal Peaks: rsd 
#> Trading Day Peaks: sa irr 
#> Overall Index of Quality of SA
#> (Acceptance Region from 0 to 1)
#> Q: 0.26 
#> Number of M statistics outside the limits: 0 
#> 
#> SA decomposition: multiplicative 
#> Seasonal moving average used for the final iteration: 
#> 3x3 (Based on the size of the global moving seasonality ratio (msr))
#> Moving average used to estimate the final trend-cycle: 9-term Henderson filter
#> 
#> ---------------------------  RUN   1   ----------------------------------------
#> 
#> 	Time Series
#> 
#> Frequency: 12 
#> Span: 1st month,1949 to 12th month,1960 
#> 
#> 	Model Definition
#> 
#> ARIMA Model: (0 1 1)(0 1 1) (Automatic Model Choice)
#> Model Span: 1st month,1949 to 12th month,1960 
#> Transformation: Automatic selection : Log(y) 
#> Regression Model: none 
#> 
#> 	Outlier Detection
#> 
#> Outlier Span: 1st month,1949 to 12th month,1960 
#> Critical |t| for outliers:	
#> aocrit1 aocrit2 lscrit1 lscrit2 tccrit1 tccrit2 
#>  "3.89"     "*"  "3.89"     "*"  "3.89"     "*" 
#> Total Number of Outliers: 0 
#> Automatically Identified Outliers: 0 
#> 
#> 	Seasonal Adjustment
#> 
#> Identifiable Seasonality: yes 
#> Seasonal Peaks: rsd 
#> Trading Day Peaks: sa irr 
#> Overall Index of Quality of SA
#> (Acceptance Region from 0 to 1)
#> Q: 0.26 
#> Number of M statistics outside the limits: 0 
#> 
#> SA decomposition: multiplicative 
#> Seasonal moving average used for the final iteration: 
#> 3x3 (Based on the size of the global moving seasonality ratio (msr))
#> Moving average used to estimate the final trend-cycle: 9-term Henderson filter
#the x12Parameter and x12Output of the x12Single is set to the previous run of x12
Ap=prev(AirPassengersX12)
summary(AirPassengersX12,oldOutput=10)
#> --------------------------   AirPassengers   ------------------------------------
#> -----------------------------------------------------------------------------------
#> 
#> 	Time Series
#> 
#> Frequency: 12 
#> Span: 1st month,1949 to 12th month,1960 
#> 
#> 	Model Definition
#> 
#> ARIMA Model: (1,1,0)(0,1,1) 
#> Model Span: 1st month,1949 to 12th month,1960 
#> Transformation: Automatic selection : Log(y) 
#> Regression Model: none 
#> 
#> 	Outlier Detection
#> 
#> Outlier Span: 1st month,1949 to 12th month,1960 
#> Critical |t| for outliers:	
#> aocrit1 aocrit2 lscrit1 lscrit2 tccrit1 tccrit2 
#>  "3.89"     "*"  "3.89"     "*"  "3.89"     "*" 
#> Total Number of Outliers: 0 
#> Automatically Identified Outliers: 0 
#> 
#> 	Seasonal Adjustment
#> 
#> Identifiable Seasonality: yes 
#> Seasonal Peaks: rsd 
#> Trading Day Peaks: sa irr 
#> Overall Index of Quality of SA
#> (Acceptance Region from 0 to 1)
#> Q: 0.26 
#> Number of M statistics outside the limits: 0 
#> 
#> SA decomposition: multiplicative 
#> Seasonal moving average used for the final iteration: 
#> 3x3 (Based on the size of the global moving seasonality ratio (msr))
#> Moving average used to estimate the final trend-cycle: 9-term Henderson filter
#> 
#> ---------------------------  RUN   1   ----------------------------------------
#> 
#> 	Time Series
#> 
#> Frequency: 12 
#> Span: 1st month,1949 to 12th month,1960 
#> 
#> 	Model Definition
#> 
#> ARIMA Model: (0 1 1)(0 1 1) (Automatic Model Choice)
#> Model Span: 1st month,1949 to 12th month,1960 
#> Transformation: Automatic selection : Log(y) 
#> Regression Model: none 
#> 
#> 	Outlier Detection
#> 
#> Outlier Span: 1st month,1949 to 12th month,1960 
#> Critical |t| for outliers:	
#> aocrit1 aocrit2 lscrit1 lscrit2 tccrit1 tccrit2 
#>  "3.89"     "*"  "3.89"     "*"  "3.89"     "*" 
#> Total Number of Outliers: 0 
#> Automatically Identified Outliers: 0 
#> 
#> 	Seasonal Adjustment
#> 
#> Identifiable Seasonality: yes 
#> Seasonal Peaks: rsd 
#> Trading Day Peaks: sa irr 
#> Overall Index of Quality of SA
#> (Acceptance Region from 0 to 1)
#> Q: 0.26 
#> Number of M statistics outside the limits: 0 
#> 
#> SA decomposition: multiplicative 
#> Seasonal moving average used for the final iteration: 
#> 3x3 (Based on the size of the global moving seasonality ratio (msr))
#> Moving average used to estimate the final trend-cycle: 9-term Henderson filter