Diagnostics summary for objects of class x12work

Diagnostics summary for objects of class "x12work".

# S3 method for class 'x12work'
summary(object,fullSummary=FALSE, spectra.detail=FALSE,almostout=FALSE,
  rsd.autocorr=NULL,quality.stat=FALSE,likelihood.stat=FALSE,aape=FALSE,id.rsdseas=FALSE,
  slidingspans=FALSE,history=FALSE,identify=FALSE,...)

Arguments

object

an object of class "x12work".

fullSummary

logical defining whether all available optional diagnostics below should be included in the summary.

spectra.detail

logical defining whether more detail on the spectra should be returned.

almostout

logical defining whether "almost" outliers should be returned.

rsd.autocorr

character or character vector specifying the type of autocorrelation of the residuals that should be returned, i.e. the autocorrelations and/or partial autocorrelations of the residuals and/or the autocorrelations of the squared residuals ("acf", "pacf", "acf2").

quality.stat

logical defining whether the second Q statistic, i.e. the Q Statistic computed w/o the M2 Quality Control Statistic, and the M statistics for monitoring and quality assessment should be returned as well.

likelihood.stat

if TRUE, the likelihood statistics AIC, AICC, BIC and HQ are returned as well as the estimated maximum value of the log likelihood function of the model for the untransformed data.

aape

logical defining whether the average absolute percentage error for forecasts should be returned.

id.rsdseas

logical defining whether the presence/absence of residual seasonality should be indicated.

slidingspans

logical defining whether the diagnostics output of the slidingspans analysis should be returned.

history

logical defining whether the diagnostics output of the (revision) history analysis should be returned.

identify

logical defining whether the (partial) autocorrelations of the residuals generated by the "identify" specification should be returned.

...

ignored at the moment

Details

Delivers a diagnostics summary.

Author

Alexander Kowarik, Angelika Meraner

See also

Examples

  data(AirPassengers)
if (FALSE) { # \dontrun{
summary(x12work(AirPassengers,...),quality.stat=TRUE,res.autocorr="acf")} # }