Class "x12Output"

Output class for x12.

Objects from the Class

Objects can be created by calls of the form new("x12Output", ...).

Slots

a1:

Object of class "ts" - the original time series.

d10:

Object of class "ts" - the final seasonal factors.

d11:

Object of class "ts" - the final seasonally adjusted data.

d12:

Object of class "ts" - the final trend cycle.

d13:

Object of class "ts" - the final irregular components.

d16:

Object of class "ts" - the combined adjustment factors.

c17:

Object of class "ts" - the final weights for the irregular component.

d9:

Object of class "ts" - the final replacements for the SI ratios.

e2:

Object of class "ts" - the differenced, transformed, seasonally adjusted data.

d8:

Object of class "ts" - the final unmodified SI ratios.

b1:

Object of class "ts" - the prior adjusted original series.

td:

Object of class "tsOrNULL" - the trading day component

otl:

Object of class "tsOrNULL" - the outlier regression series

sp0:

Object of class "spectrum" - the spectrum of the original series.

sp1:

Object of class "spectrum" - the spectrum of the differenced seasonally adjusted series.

sp2:

Object of class "spectrum" - the spectrum of modified irregular series.

spr:

Object of class "spectrum" - the spectrum of the regARIMA model residuals.

forecast:

Object of class "fbcast" - the point forecasts with prediction intervals

backcast:

Object of class "fbcast" - the point backcasts with prediction intervals

dg:

Object of class "list", containing several seasonal adjustment and regARIMA modeling diagnostics, i.e.:
x11regress, transform, samode, seasonalma, trendma, arimamdl, automdl, regmdl, nout, nautoout, nalmostout, almostoutlier, crit, outlier, userdefined, autooutlier, peaks.seas, peaks.td, id.seas, id.rsdseas, spcrsd, spcori, spcsa, spcirr, m1, m2, m3, m4, m5, m6, m7, m8, m9, m10, m11, q, q2, nmfail, loglikelihood, aic, aicc, bic, hq, aape, autotransform, ifout, rsd.acf, rsd.pacf, rsd.acf2, tsName, frequency, span,...

% \item{\code{seats}:}{Object of class \code{"logical"} ~~ }
file:

Object of class "character" - path to the output directory and filename

tblnames:

Object of class "character" - tables read into R

Rtblnames:

Object of class "character" - names of tables read into R

Methods

summary

signature(object = "x12Output"): ...

plot

signature(object = "x12Output"): ...

plotSpec

signature(object = "x12Output"): ...

plotSeasFac

signature(object = "x12Output"): ...

plotRsdAcf

signature(object = "x12Output"): ...

Author

Alexander Kowarik, Angelika Meraner

See also

Examples

data(AirPassengersX12)
summary(AirPassengersX12)
#> --------------------------   AirPassengers   ------------------------------------
#> -----------------------------------------------------------------------------------
#> 
#> 	Time Series
#> 
#> Frequency: 12 
#> Span: 1st month,1949 to 12th month,1960 
#> 
#> 	Model Definition
#> 
#> ARIMA Model: (1,1,0)(0,1,1) 
#> Model Span: 1st month,1949 to 12th month,1960 
#> Transformation: Automatic selection : Log(y) 
#> Regression Model: none 
#> 
#> 	Outlier Detection
#> 
#> Outlier Span: 1st month,1949 to 12th month,1960 
#> Critical |t| for outliers:	
#> aocrit1 aocrit2 lscrit1 lscrit2 tccrit1 tccrit2 
#>  "3.89"     "*"  "3.89"     "*"  "3.89"     "*" 
#> Total Number of Outliers: 0 
#> Automatically Identified Outliers: 0 
#> 
#> 	Seasonal Adjustment
#> 
#> Identifiable Seasonality: yes 
#> Seasonal Peaks: rsd 
#> Trading Day Peaks: sa irr 
#> Overall Index of Quality of SA
#> (Acceptance Region from 0 to 1)
#> Q: 0.26 
#> Number of M statistics outside the limits: 0 
#> 
#> SA decomposition: multiplicative 
#> Seasonal moving average used for the final iteration: 
#> 3x3 (Based on the size of the global moving seasonality ratio (msr))
#> Moving average used to estimate the final trend-cycle: 9-term Henderson filter
showClass("x12Output")
#> Class "x12Output" [package "x12"]
#> 
#> Slots:
#>                                                                             
#> Name:         a1       d10       d11       d12       d13       d16       c17
#> Class:        ts        ts        ts        ts        ts        ts        ts
#>                                                                             
#> Name:         d9        e2        d8        b1        td       otl       sp0
#> Class:        ts        ts        ts        ts  tsOrNULL  tsOrNULL  spectrum
#>                                                                             
#> Name:        sp1       sp2       spr  forecast  backcast        dg      file
#> Class:  spectrum  spectrum  spectrum    fbcast    fbcast      list character
#>                           
#> Name:   tblnames Rtblnames
#> Class: character character