Class "x12Parameter"

Parameter class for x12.

Objects from the Class

Objects can be created by calls of the form new("x12Parameter", ...).

Slots

%\item{\code{period}:}{Object of class \code{"numeric"} ~~ }
series.span:

Object of class "numericOrNULLOrcharacter" - vector of length 4, limiting the data used for the calculations and analysis to a certain time interval.
Start and end date of said time interval can be specified by 4 integers in the format c(start year, start seasonal period, end year, end seasonal period) If the start or end date of the time series object should be used, the respective year and seasonal period are to be set to NA.

series.modelspan:

Object of class "numericOrNULLOrcharacter" - vector of length 4, defining the start and end date of the time interval of the data that should be used to determine all regARIMA model coefficients. Specified in the same way as span.

%\item{\code{decimals}:}{Object of class \code{"numeric"} - number of decimal places of the data.}
transform.function:

Object of class "character" - transform parameter for x12 ("auto", "log", "none").

transform.power:

Object of class "numericOrNULL" - numeric value specifying the power of the Box Cox power transformation.

transform.adjust:

Object of class "characterOrNULL" - determines the type of adjustment to be performed, i.e. transform.adjust="lom" for length-of-month adjustment on monthly data, transform.adjust="loq" for length-of-quarter adjustment on quarterly data or transform.adjust="lpyear" for leap year adjustment of monthly or quarterly data (which is only allowed when either transform.power=0 or transform.function="log").

regression.variables:

Object of class "characterOrNULL" - character or character vector representing the names of the regression variables.

regression.user:

Object of class "characterOrNULL" - character or character vector defining the user parameters in the regression argument.

regression.file:

Object of class "characterOrNULL" - path to the file containing the data values of all regression.user variables.

regression.usertype:

Object of class "characterOrNULL" - character or character vector assigning a type of model-estimated regression effect on each user parameter in the regression argument ("seasonal", "td", "lpyear", "user", ...).
By specifying a character vector of length greater one each variable can be given its own type. Otherwise the same type will be used for all user parameters.

regression.centeruser:

Object of class "characterOrNULL" - character specifying the removal of the (sample) mean or the seasonal means from the user parameters in the regression argument ("mean", "seasonal").
Default is no modification of the respective user-defined regressors.

regression.start:

Object of class "numericOrNULLOrcharacter" - start date for the values of the regression.user variables, specified as a vector of two integers in the format c(year, seasonal period).

regression.aictest:

Object of class "characterOrNULL" - character vector defining the regression variables for which an AIC test is to be performed.

% \item{\code{outlier}:}{Object of class \code{"logical"} - if \code{TRUE}, the "outlier" specification will be enabled in the spc file.}
outlier.types:

Object of class "characterOrNULL" - to enable the "outlier" specification in the spc file, this parameter has to be defined by a character or character vector determining the method(s) used for outlier detection ("AO", "LS", "TC", "all").

outlier.critical:

Object of class "listOrNULLOrnumeric" - number specifying the critical value used for outlier detection (same value used for all types of outliers) or named list (possible names of list elements being AO,LS and TC) where each list element specifies the respective critical value used for detecting the corresponding type of outlier.
If not specified, the default critical value is used.

outlier.span:

Object of class "numericOrNULLOrcharacter" - vector of length 4, defining the span for outlier detection. Specified in the same way as span.

outlier.method:

Object of class "characterOrNULL" - character determining how detected outliers should be added to the model ("addone", "addall"). If not specified,"addone" is used by default.

% (For x11 regression this argument cannot be used if the sigma argument is used.)
identify:

Object of class "logical" - if TRUE, the "identify" specification will be enabled in the spc file.

identify.diff:

Object of class "numericOrNULL" - number or vector representing the orders of nonseasonal differences specified, default is 0.

identify.sdiff:

Object of class "numericOrNULL" - number or vector representing the orders of seasonal differences specified, default is 0.

identify.maxlag:

Object of class "numericOrNULL" - number of lags specified for the ACFs and PACFs, default is 36 for monthly series and 12 for quarterly series.

arima.model:

Object of class "numericOrNULL" - vector of length 3, defining the arima parameters.

arima.smodel:

Object of class "numericOrNULL" - vector of length 3, defining the sarima parameters.

arima.ar:

Object of class "numericOrNULLOrcharacter" - numeric or character vector specifying the initial values for nonseasonal and seasonal autoregressive parameters in the order that they appear in the arima.model argument. Empty positions are created with NA.

arima.ma:

Object of class "numericOrNULLOrcharacter" - numeric or character vector specifying the initial values for all moving average parameters in the order that they appear in the arima.model argument. Empty positions are created with NA.

automdl:

Object of class "logical" - TRUE/FALSE for activating auto modeling.

automdl.acceptdefault:

Object of class "logical" - logical for automdl defining whether the default model should be chosen if the Ljung-Box Q statistic for its model residuals is acceptable.

automdl.balanced:

Object of class "logical" - logical for automdl defining whether the automatic model procedure will tend towards balanced models. TRUE yields the same preference as the TRAMO program.

automdl.maxorder:

Object of class "numeric" - vector of length 2, specifying the maximum order for automdl. Empty positions are created with NA.

automdl.maxdiff:

Object of class "numeric" - vector of length 2, specifying the maximum diff. order for automdl. Empty positions are created with NA.

forecast_years:

Object of class "numericOrNULL" - number of years to forecast, default is 1 year.

backcast_years:

Object of class "numericOrNULL" - number of years to backcast, default is no backcasts.

forecast_conf:

Object of class "numeric" - probability for the confidence interval of forecasts.

forecast_save:

Object of class "character" either "ftr"(in transformed scaling) or "fct"(in original scaling)

estimate:

Object of class "logical" - if TRUE, the term "estimate" will be added to the spc file.

estimate.outofsample:

Object of class "logical" - logical defining whether "out of sample" or "within sample" forecast errors should be used in calculating the average magnitude of forecast errors over the last three years.

check:

Object of class "logical" - TRUE/FALSE for activating the "check" specification in the spc file.

check.maxlag:

Object of class "numericOrNULL" - the number of lags requested for the residual sample ACF and PACF, default is 24 for monthly series and 8 for quarterly series.

slidingspans:

Object of class "logical" - if TRUE, "slidingspans" specification will be enabled in the spc file.

slidingspans.fixmdl:

Object of class "characterOrNULL" - ("yes" (default), "no", "clear").

slidingspans.fixreg:

Object of class "characterOrNULL" - character or character vector specifying the trading day, holiday, outlier or other user-defined regression effects to be fixed ("td", "holiday", "outlier", "user"). All other regression coefficients will be re-estimated for each sliding span.

slidingspans.length:

Object of class "numericOrNULL" - numeric value specifying the length of each span in months or quarters (>3 years, <17 years).

slidingspans.numspans:

Object of class "numericOrNULL" - numeric value specifying the number of sliding spans used to generate output for comparisons (must be between 2 and 4, inclusive).

slidingspans.outlier:

Object of class "characterOrNULL" - ("keep" (default), "remove", "yes").

slidingspans.additivesa:

Object of class "characterOrNULL" - ("difference" (default), "percent").

slidingspans.start:

Object of class "numericOrNULLOrcharacter" - specified as a vector of two integers in the format c(start year, start seasonal period).

history:

if TRUE, the history specification will be enabled.

history.estimates:

Object of class "characterOrNULL" - character or character vector determining which estimates from the regARIMA modeling and/or the x11 seasonal adjustment will be analyzed in the history analysis ("sadj" (default), "sadjchng", "trend", "trendchng", "seasonal", "aic", "fcst").

history.fixmdl:

Object of class "logical" - logical determining whether the regARIMA model will be re-estimated during the history analysis.

history.fixreg:

Object of class "characterOrNULL" - character or character vector specifying the trading day, holiday, outlier or other user-defined regression effects to be fixed ("td", "holiday", "outlier", "user"). All other coefficients will be re-estimated for each history span.

history.outlier:

Object of class "characterOrNULL" - ("keep" (default), "remove", "auto")

history.sadjlags:

Object of class "numericOrNULL" - integer or vector specifying up to 5 revision lags (each >0) that will be analyzed in the revisions analysis of lagged seasonal adjustments.

history.trendlags:

Object of class "numericOrNULL" - integer or vector specifying up to 5 revision lags (each >0) that will be used in the revision history of the lagged trend components.

history.start:

Object of class "numericOrNULLOrcharacter" - specified as a vector of two integers in the format c(start year, start seasonal period).

history.target:

Object of class "characterOrNULL" - character determining whether the revisions of the seasonal adjustments and trends calculated at the lags specified in history.sadjlags and history.trendlags should be defined by the deviation from the concurrent estimate or the deviation from the final estimate ("final" (default), "concurrent").

x11.sigmalim:

Object of class "numericOrNULL" - vector of length 2, defining the limits for sigma in the x11 methodology, used to downweight extreme irregular values in the internal seasonal adjustment iterations.

x11.type:

Object of class "characterOrNULL" - character, i.e. "summary", "trend" or "sa". If x11.type="trend", x11 will only be used to estimate the final trend-cycle as well as the irregular components and to adjust according to trading days. The default setting is type="sa" where a seasonal decomposition of the series is calculated.

% \item{\code{x11.type}:}{Object of class "logical" - if \code{TRUE}, x11 will only be used to estimate trend and to adjust according to trading days.}
x11.sfshort:

Object of class "logical" - logical controlling the seasonal filter to be used if the series is at most 5 years long. If TRUE, the arguments of the seasonalma filter will be used wherever possible. If FALSE, a stable seasonal filter will be used irrespective of seasonalma.

x11.samode:

Object of class "characterOrNULL" - character defining the type of seasonal adjustment decomposition calculated ("mult", "add", "pseudoadd", "logadd").

x11.seasonalma:

Object of class "characterOrNULL" - character or character vector of the format c("snxm","snxm", ...) defining which seasonal nxm moving average(s) should be used for which calendar months or quarters to estimate the seasonal factors. If only one ma is specified, the same ma will be used for all months or quarters. If not specified, the program will invoke an automatic choice.

x11.trendma:

Object of class "numericOrNULL" - integer defining the type of Henderson moving average used for estimating the final trend cycle. If not specified, the program will invoke an automatic choice.

x11.appendfcst:

Object of class "logical" - logical defining whether forecasts should be included in certain x11 tables.

x11.appendbcst:

Object of class "logical" - logical defining whether forecasts should be included in certain x11 tables.

x11.calendarsigma:

Object of class "characterOrNULL" - regulates the way the standard errors used for the detection and adjustment of extreme values should be computed ("all", "signif", "select" or no specification).

x11.excludefcst:

Object of class "logical" - logical defining if forecasts and backcasts from the regARIMA model should not be used in the generation of extreme values in the seasonal adjustment routines.

x11.final:

Object of class "character" - character or character vector specifying which type(s) of prior adjustment factors should be removed from the final seasonally adjusted series ("AO", "LS", "TC", "user", "none").

x11regression:

Object of class "logical" - if TRUE, x11Regression will be performed using the respective regression and outlier commands above, i.e. regression.variables, regression.user, regression.file, regression.usertype, regression.centeruser and regression.start as well as outlier.critical, outlier.span and outlier.method.

% \item{\code{seats}:}{Object of class \code{"logical"} - \code{TRUE}/\code{FALSE} for activating SEATS, at the moment SEATS is not properly implemented.} % \item{\code{seatsparameter}:}{Object of class \code{"characterOrNULL"} - string defining the seats parameter.}

Methods

getP

signature(object = "x12Parameter"): ...

setP

signature(object = "x12Parameter"): ...

Author

Alexander Kowarik, Angelika Meraner

Examples

showClass("x12Parameter")
#> Class "x12Parameter" [package "x12"]
#> 
#> Slots:
#>                                                         
#> Name:               series.span         series.modelspan
#> Class: numericOrNULLOrcharacter numericOrNULLOrcharacter
#>                                                         
#> Name:        transform.function          transform.power
#> Class:                character            numericOrNULL
#>                                                         
#> Name:          transform.adjust     regression.variables
#> Class:          characterOrNULL          characterOrNULL
#>                                                         
#> Name:           regression.user          regression.file
#> Class:          characterOrNULL          characterOrNULL
#>                                                         
#> Name:       regression.usertype    regression.centeruser
#> Class:          characterOrNULL          characterOrNULL
#>                                                         
#> Name:          regression.start       regression.aictest
#> Class: numericOrNULLOrcharacter          characterOrNULL
#>                                                         
#> Name:             outlier.types         outlier.critical
#> Class:          characterOrNULL      listOrNULLOrnumeric
#>                                                         
#> Name:              outlier.span           outlier.method
#> Class: numericOrNULLOrcharacter          characterOrNULL
#>                                                         
#> Name:                  identify            identify.diff
#> Class:                  logical            numericOrNULL
#>                                                         
#> Name:            identify.sdiff          identify.maxlag
#> Class:            numericOrNULL            numericOrNULL
#>                                                         
#> Name:               arima.model             arima.smodel
#> Class:            numericOrNULL            numericOrNULL
#>                                                         
#> Name:                  arima.ar                 arima.ma
#> Class: numericOrNULLOrcharacter numericOrNULLOrcharacter
#>                                                         
#> Name:                   automdl    automdl.acceptdefault
#> Class:                  logical                  logical
#>                                                         
#> Name:          automdl.balanced         automdl.maxorder
#> Class:                  logical                  numeric
#>                                                         
#> Name:           automdl.maxdiff           forecast_years
#> Class:                  numeric            numericOrNULL
#>                                                         
#> Name:            backcast_years            forecast_conf
#> Class:            numericOrNULL                  numeric
#>                                                         
#> Name:             forecast_save                 estimate
#> Class:                character                  logical
#>                                                         
#> Name:      estimate.outofsample                    check
#> Class:                  logical                  logical
#>                                                         
#> Name:              check.maxlag             slidingspans
#> Class:            numericOrNULL                  logical
#>                                                         
#> Name:       slidingspans.fixmdl      slidingspans.fixreg
#> Class:          characterOrNULL          characterOrNULL
#>                                                         
#> Name:       slidingspans.length    slidingspans.numspans
#> Class:            numericOrNULL            numericOrNULL
#>                                                         
#> Name:      slidingspans.outlier  slidingspans.additivesa
#> Class:          characterOrNULL          characterOrNULL
#>                                                         
#> Name:        slidingspans.start                  history
#> Class: numericOrNULLOrcharacter                  logical
#>                                                         
#> Name:         history.estimates           history.fixmdl
#> Class:          characterOrNULL                  logical
#>                                                         
#> Name:            history.fixreg          history.outlier
#> Class:          characterOrNULL          characterOrNULL
#>                                                         
#> Name:          history.sadjlags        history.trendlags
#> Class:            numericOrNULL            numericOrNULL
#>                                                         
#> Name:             history.start           history.target
#> Class: numericOrNULLOrcharacter          characterOrNULL
#>                                                         
#> Name:              x11.sigmalim                 x11.type
#> Class:            numericOrNULL          characterOrNULL
#>                                                         
#> Name:               x11.sfshort               x11.samode
#> Class:                  logical          characterOrNULL
#>                                                         
#> Name:            x11.seasonalma              x11.trendma
#> Class:          characterOrNULL            numericOrNULL
#>                                                         
#> Name:            x11.appendfcst           x11.appendbcst
#> Class:                  logical                  logical
#>                                                         
#> Name:         x11.calendarsigma          x11.excludefcst
#> Class:          characterOrNULL                  logical
#>                                                         
#> Name:                 x11.final            x11regression
#> Class:                character                  logical